requirements.txt
python-telegram-bot>=20.3
requests>=2.31.0
python-dotenv>=1.0.0
groq>=0.4.0
sentry-sdk>=1.40.0
stripe>=5.4.0
flask>=2.3.0